fixup: move classloader instantiation to injectable ClassLoaderFactory
