+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/isolation/Plugins.java
@@ -38,8 +38,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.file.Path;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -59,23 +57,16 @@ public class Plugins {
     private final PluginScanResult scanResult;
 
     public Plugins(Map<String, String> props) {
-        this(props, Plugins.class.getClassLoader());
+        this(props, Plugins.class.getClassLoader(), new ClassLoaderFactory());
     }
 
     // VisibleForTesting
-    Plugins(Map<String, String> props, ClassLoader parent) {
+    Plugins(Map<String, String> props, ClassLoader parent, ClassLoaderFactory factory) {
         String pluginPath = WorkerConfig.pluginPath(props);
         List<Path> pluginLocations = PluginUtils.pluginLocations(pluginPath);
-        delegatingLoader = newDelegatingClassLoader(parent);
-        Set<PluginSource> pluginSources = DelegatingClassLoader.sources(pluginLocations, delegatingLoader);
+        delegatingLoader = factory.newDelegatingClassLoader(parent);
+        Set<PluginSource> pluginSources = DelegatingClassLoader.sources(pluginLocations, delegatingLoader, factory);
         scanResult = initLoaders(pluginSources);
-    }
-
-    // VisibleForTesting
-    protected DelegatingClassLoader newDelegatingClassLoader(ClassLoader parent) {
-        return AccessController.doPrivileged(
-                (PrivilegedAction<DelegatingClassLoader>) () -> new DelegatingClassLoader(parent)
-        );
     }
 
     public PluginScanResult initLoaders(Set<PluginSource> pluginSources) {
