+++ connect/runtime/src/test/java/org/apache/kafka/connect/runtime/isolation/DelegatingClassLoaderTest.java
@@ -112,8 +112,9 @@ public class DelegatingClassLoaderTest {
     }
 
     private DelegatingClassLoader initClassLoader(List<Path> pluginLocations) {
-        DelegatingClassLoader classLoader = new DelegatingClassLoader();
-        Set<PluginSource> pluginSources = DelegatingClassLoader.sources(pluginLocations, classLoader);
+        ClassLoaderFactory factory = new ClassLoaderFactory();
+        DelegatingClassLoader classLoader = factory.newDelegatingClassLoader(DelegatingClassLoader.class.getClassLoader());
+        Set<PluginSource> pluginSources = DelegatingClassLoader.sources(pluginLocations, classLoader, factory);
         PluginScanResult scanResult = new ReflectionScanner().discoverPlugins(pluginSources);
         classLoader.installDiscoveredPlugins(scanResult);
         return classLoader;
